Add detection of additional weak and insecure TLS options
Add unit tests and code to detect the insecure options "ssl_protocols SSLv2" and "ssl_protocols SSLv3", and the weak option "ssl_early_data on;" which enables the zero round-trip time (0-RTT) mode in TLS 1.3.
Remove unit tests for the reporter functions because, for the time being, these function merely print arrays $tls_audit_weak and $tls_audit_insecure to the command line (so, there is nothing to test really).